requirements/requirements.txt
psycopg2==2.8.2
alembic==1.0.10
sqlalchemy==1.2.11
flask_cache==0.13.1
flask_wtf==0.14.2
flask_restful==0.3.7
flask_httpauth==3.3.0
gunicorn==19.9.0
flask_login==0.4.1
bcrypt==3.1.6
pillow==5.2.0
